oltmonitor/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==5.2.12
djangorestframework==3.16.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
ply==3.11
pyasn1==0.4.8
pycryptodomex==3.23.0
pysmi==1.6.3
pysnmp==4.4.12
python-decouple==3.8
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
